+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#new)
     * Added support for admin reporting available in Grav 1.6
+1. [](#bugfix)
+    * Fixed issue with twig auto-escaping
 
 # v2.0.0
 ## 09/30/2018