+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Added support for admin reporting available in Grav 1.6
 1. [](#bugfix)
     * Fixed issue with twig auto-escaping
+    * Fixed problems plugin potentially breaking CLI command if plugins get initialized
 
 # v2.0.0
 ## 09/30/2018